backend/requirements.txt
dj_database_url==0.5.0
Django==3.2.5
django_polymorphic==3.1.0
django_rest_polymorphic==0.1.9
djangorestframework==3.13.1
django-cors-headers==3.11.0
psycopg2
gunicorn
whitenoise